backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
firebase-admin
librosa
scikit-learn
numpy
scipy
soundfile
python-dotenv
